feat: extend parseRefString heuristic to allow more CIs
Previously, local / git refs were estimated using such heuristic:
- was being considered as remote any ref with no version specified AND ref starting with "github.com"

However, as we are now extending the logic to other CI providers, this
needs to be updated.

The current logic is extended:
- stat if the dir is present
- if it is not, we then try to estimate the root of the remote (repoRootForImportPath)
- If it fails, we fallback as a local remote (same behavior as it was in the past)

The main reason to fallback as a local remote is that a lot of `dagger install` / `dagger call` design decisions have been implemented, and is not part of the scope of this PR. e.g. `dagger init --source=new_dir` if new_dir is not present

As part of this commit, we also implement splitRootAndSubdir(), which extracts the root of the repo and the subdir from the ref.

Supported refs at the moment:
- Public CIs: github, gitlab, gitea, bitbucket
- vanity URLs

- Path to private refs support (especially with GitLab)

Signed-off-by: grouville <guillaume@dagger.io>