chore: rework builtinOrInput to inputType
We were previously doing some fancy reflection to populate the values:
but actually it's totally valid for these values to be completely empty
(I think).

We're only ever using this type value to decode other inputs, and
inspect the type - we're never using it's value, so we should actually
discard it.

Found this while attempting to rework the signature of `DecodeInput`,
we just don't need to be calling it here.

Signed-off-by: Justin Chadwell <me@jedevc.com>